improvement: update k7s

+++ k7s/deploy.sh
@@ -4,8 +4,11 @@ if [ "$1" == "bash" ]; then
     exec /bin/bash
 fi
 
-[ -f "/root/.sealos/kube.tgz" ] && rm -rf /root/.sealos/kube.tgz || mkdir -p /root/.sealos
+if [ "$1" == "init" ]; then
+    [ -f "/root/.sealos/kube.tgz" ] && rm -rf /root/.sealos/kube.tgz || mkdir -p /root/.sealos
+    cp -a /kube.tgz /root/.sealos/kube.tgz
+    exec sealos init --repo registry.cn-beijing.aliyuncs.com/k7scn --version 1.19.2 --pkg-url /root/.sealos/kube.tgz ${@:2}
+else
+    exec sealos join ${@:2}
+fi
 
-cp -a /kube.tgz /root/.sealos/kube.tgz
-
-exec sealos init --repo registry.cn-beijing.aliyuncs.com/k7scn --version 1.19.2 --pkg-url /root/.sealos/kube.tgz $@
